stable

+++ TelegramMultiBot.Database/DTO/JobResultInfo.cs
@@ -5,6 +5,6 @@ public class JobResultInfo
     public string Id { get; set; }
     public long Seed { get; set; }
     public string Info { get; set; }
-    public long RenderTime { get; set; }
+    public double RenderTime { get; set; }
     public string FilePath { get; set; }
 }
